ci(analysis): gate on clang-tidy with a tuned check set
ESP-IDF 的默认 clang-tidy 集合在本项目上产生 2137 条告警，其中 130 条
bugprone-branch-clone 全部来自 ESP_LOGx 展开成的条件链，46 条
cognitive-complexity 针对的是状态机与协议解析这类天然多分支的函数——
真正的问题会被这些噪声淹掉，这样的报告没人会看。

tools/clang-check.sh 从 `-*` 起步，只挑 clang-analyzer / bugprone / performance，
再排除两项高误报的检查，把我们自己代码的诊断压到 0 条，然后作为硬门禁进 CI。

检查集合只能从命令行传：写在 .clang-tidy 里会被 idf.py clang-check 自己传的
-checks 覆盖掉，所以脚本里带注释说明了这一点。

它当场发现并修掉了两处真实的窄化转换：

- OTA 上传循环把 size_t 的 content_len 塞进 int（当前被分区容量上限挡着，
  但那是巧合，不是设计）
- httpd_resp_send_chunk 的 size_t → ssize_t

顺带把 uptime 的整数除法改成浮点域换算，意图更明确。

另外补齐文档：README 的项目结构、Kconfig 菜单、访问令牌用法与找回途径、
固件签名的开启方式；AGENTS.md 补上静态分析与 menuconfig 的入口。

+++ components/kenko_web/api_handlers.c
@@ -340,7 +340,8 @@ static esp_err_t system_info_handler(httpd_req_t *req)
     cJSON_AddItemToObject(root, "firmware", firmware);
 
     cJSON *runtime = cJSON_CreateObject();
-    cJSON_AddNumberToObject(runtime, "uptime_ms", (double)(esp_timer_get_time() / 1000));
+    /* esp_timer_get_time() 返回微秒，这里在浮点域上换算成毫秒。 */
+    cJSON_AddNumberToObject(runtime, "uptime_ms", (double)esp_timer_get_time() / 1000.0);
 
     cJSON *heap = cJSON_CreateObject();
     cJSON_AddItemToObject(heap, "internal", build_heap_json(HEAP_CAPS_INTERNAL));
@@ -547,10 +548,10 @@ static esp_err_t ota_upload(httpd_req_t *req)
         return send_error(req, 500, "no_memory", "out of memory");
     }
 
-    int remaining = req->content_len;
+    size_t remaining = (size_t)req->content_len;
     while (remaining > 0) {
-        int wanted = remaining < OTA_CHUNK_BYTES ? remaining : OTA_CHUNK_BYTES;
-        int received = httpd_req_recv(req, buffer, (size_t)wanted);
+        size_t wanted = remaining < OTA_CHUNK_BYTES ? remaining : (size_t)OTA_CHUNK_BYTES;
+        int received = httpd_req_recv(req, buffer, wanted);
         if (received == HTTPD_SOCK_ERR_TIMEOUT) {
             continue;
         }
@@ -567,7 +568,7 @@ static esp_err_t ota_upload(httpd_req_t *req)
             ota_service_get_status(&status);
             return send_error(req, 400, "image_rejected", status.message);
         }
-        remaining -= received;
+        remaining -= (size_t)received;
     }
     free(buffer);
